ボタンのイベントを定義

+++ Assets/AxisOrange/AxisOrangeData.cs
@@ -4,12 +4,12 @@ using UnityEngine;
 
 namespace AxisOrange {
     public readonly struct AxisOrangeData {
-        public readonly long timestamp;
+        public readonly uint timestamp;
         public readonly Vector3 acc;
         public readonly Vector3 gyro;
         public readonly Quaternion quaternion;
 
-        public AxisOrangeData(long timestamp, Vector3 acc, Vector3 gyro, Quaternion quaternion) {
+        public AxisOrangeData(uint timestamp, Vector3 acc, Vector3 gyro, Quaternion quaternion) {
             this.timestamp = timestamp;
             this.acc = acc;
             this.gyro = gyro;
